requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0
